requirements.txt
flask
chromadb
google-genertiveai
yfinance
pandas
bs4
openai
langchain_community
statsmodels
duckduckgo-search
flask-cors